Conflicts:
	Source/eu.modelwriter.writer.markers/src/eu/modelwriter/marker/action/DeleteMarkerAction.java

+++ Source/eu.modelwriter.writer.markers/src/eu/modelwriter/marker/action/DeleteAllMarkerAction.java
@@ -8,7 +8,9 @@ import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IMarker;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.emf.common.util.URI;
+import org.eclipse.emf.ecore.EModelElement;
 import org.eclipse.emf.ecore.ENamedElement;
+import org.eclipse.emf.ecore.EObject;
 import org.eclipse.emf.ecore.presentation.EcoreEditor;
 import org.eclipse.emf.ecore.util.EcoreUtil;
 import org.eclipse.jface.action.IAction;
@@ -83,20 +85,41 @@ public class DeleteAllMarkerAction implements IEditorActionDelegate {
           }
         }
       } else if (selection instanceof ITreeSelection) {
-        if (editor instanceof EcoreEditor) {
-          ITreeSelection treeSelection = (ITreeSelection) selection;
-          if (selection != null && Activator.getEditor() instanceof EcoreEditor
-              && treeSelection.getFirstElement() instanceof ENamedElement
+        ITreeSelection treeSelection = (ITreeSelection) selection;
+        if (selection != null && editor instanceof EcoreEditor) {
+          if (treeSelection.getFirstElement() instanceof ENamedElement
               && ((ENamedElement) treeSelection.getFirstElement()).getName() != null
               && !((ENamedElement) treeSelection.getFirstElement()).getName().isEmpty()) {
 
-           URI uri = EcoreUtil.getURI((ENamedElement) treeSelection.getFirstElement());
+            URI uri = EcoreUtil.getURI((ENamedElement) treeSelection.getFirstElement());
 
             beDeleted = MarkerFactory.findMarkersByUri(file, uri.toString());
             if (beDeleted != null && beDeleted.exists()) {
               updateTargets(beDeleted);
               updateSources(beDeleted);
+
+              MessageDialog dialog = new MessageDialog(Activator.getShell(),
+                  "Mark will be deleted by this wizard", null,
+                  "\"" + beDeleted.getAttribute(IMarker.TEXT)
+                      + "\" has been seleceted to be unmarked",
+                  MessageDialog.INFORMATION, new String[] {"OK"}, 0);
               beDeleted.delete();
+              dialog.open();
+            }
+          } else if (!((EObject) treeSelection.getFirstElement() instanceof EModelElement)) {
+            URI uri = EcoreUtil.getURI((EObject) treeSelection.getFirstElement());
+            beDeleted = MarkerFactory.findMarkersByUri(file, uri.toString());
+            if (beDeleted != null && beDeleted.exists()) {
+              updateTargets(beDeleted);
+              updateSources(beDeleted);
+
+              MessageDialog dialog = new MessageDialog(Activator.getShell(),
+                  "Mark will be deleted by this wizard", null,
+                  "\"" + beDeleted.getAttribute(IMarker.TEXT)
+                      + "\" has been seleceted to be unmarked",
+                  MessageDialog.INFORMATION, new String[] {"OK"}, 0);
+              beDeleted.delete();
+              dialog.open();
             }
           }
         }
